requirements.txt
Flask
psutil
speedtest-cli
wakeonlan
scapy
mac-vendor-lookup
